chore: remove unused array logic and fix flaky test
