--- api/src/__test__/core/modules/array/add-disk-to-array.test.ts
@@ -1,5 +1,0 @@
-import { test } from 'vitest';
-
-test.todo('Adds a disk to the array');
-
-test.todo('Fails to add the disk if the array is started');
