--- api/src/__test__/core/modules/array/remove-disk-from-array.test.ts
@@ -1,5 +1,0 @@
-import { test } from 'vitest';
-
-test.todo('Removes a disk from the array');
-
-test.todo('Fails to remove the disk if the array is started');
