--- api/src/__test__/core/modules/array/update-array.test.ts
@@ -1,5 +1,0 @@
-import { test } from 'vitest';
-
-test.todo('Starts the array');
-
-test.todo('Stops the array');
