--- api/src/__test__/core/modules/array/update-parity-check.test.ts
@@ -1,7 +1,0 @@
-import { test } from 'vitest';
-
-test.todo('Can start a parity check');
-
-test.todo('Can pause a parity check');
-
-test.todo('Can start a parity check');
