--- api/src/core/modules/array/index.ts
@@ -1,6 +1,0 @@
-// Created from 'create-ts-index'
-
-export * from './add-disk-to-array.js';
-export * from './remove-disk-from-array.js';
-export * from './update-array.js';
-export * from './update-parity-check.js';
