+++ api/src/unraid-api/graph/resolvers/array/array.service.ts
@@ -1,6 +1,7 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 
 import { capitalCase, constantCase } from 'change-case';
+import { GraphQLError } from 'graphql';
 
 import type { ArrayDiskInput, ArrayStateInput, ArrayType } from '@app/graphql/generated/api/types.js';
 import { AppError } from '@app/core/errors/app-error.js';
@@ -141,5 +142,46 @@ export class ArrayService {
         });
 
         return getArrayData();
+    }
+
+    async updateParityCheck({ wantedState, correct }: { wantedState: string; correct: boolean }) {
+        const { getters } = await import('@app/store/index.js');
+        const running = getters.emhttp().var.mdResync !== 0;
+        const states = {
+            pause: {
+                cmdNoCheck: 'Pause',
+            },
+            resume: {
+                cmdCheck: 'Resume',
+            },
+            cancel: {
+                cmdNoCheck: 'Cancel',
+            },
+            start: {
+                cmdCheck: 'Check',
+            },
+        };
+
+        let allowedStates = Object.keys(states);
+
+        // Only allow starting a check if there isn't already one running
+        if (running) {
+            // Remove 'start' from allowed states when a check is already running
+            allowedStates = allowedStates.filter((state) => state !== 'start');
+        }
+
+        // Only allow states from states object
+        if (!allowedStates.includes(wantedState)) {
+            throw new GraphQLError(`Invalid parity check state: ${wantedState}`);
+        }
+
+        // Should we write correction to the parity during the check
+        const writeCorrectionsToParity = wantedState === 'start' && correct;
+
+        await emcmd({
+            startState: 'STARTED',
+            ...states[wantedState],
+            ...(writeCorrectionsToParity ? { optionCorrect: 'correct' } : {}),
+        });
     }
 }
